+++ packages/rust/src/quest_settings.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct DailyQuest {
-    pub bonus_multiplier: Option<u64>,
+    pub bonus_multiplier: Option<f64>,
     pub buckets_per_day: u64,
-    pub streak_bonus_multiplier: Option<u64>,
+    pub streak_bonus_multiplier: Option<f64>,
     pub streak_length: u64,
 }
